+++ RiskOfChaos/EffectDefinitions/Character/ForceActivateRandomSkill.cs
@@ -1,5 +1,4 @@
-﻿using RiskOfChaos.EffectHandling;
-using RiskOfChaos.EffectHandling.EffectClassAttributes;
+﻿using RiskOfChaos.EffectHandling.EffectClassAttributes;
 using RiskOfChaos.EffectHandling.EffectClassAttributes.Methods;
 using RiskOfChaos.ModifierController.SkillSlots;
 using RoR2;
@@ -7,8 +6,8 @@ using System;
 
 namespace RiskOfChaos.EffectDefinitions.Character
 {
-    [ChaosEffect("force_activate_random_skill", DefaultSelectionWeight = 0.3f, EffectWeightReductionPercentagePerActivation = 80f)]
-    [ChaosTimedEffect(TimedEffectType.UntilStageEnd)]
+    [ChaosEffect("force_activate_random_skill", DefaultSelectionWeight = 0.6f)]
+    [ChaosTimedEffect(90f)]
     [EffectConfigBackwardsCompatibility("Effect: Force Activate Random Skill (Lasts 1 stage)")]
     public sealed class ForceActivateRandomSkill : TimedEffect, ISkillSlotModificationProvider
     {